+++ clang-tools-extra/clang-tidy/hooks/ReserveLimitCheck.h
@@ -1,0 +1,37 @@
+//===--- ReserveLimitCheck.h - clang-tidy -----------------------*- C++ -*-===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_HOOKS_RESERVELIMITCHECK_H
+#define LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_HOOKS_RESERVELIMITCHECK_H
+
+#include "../ClangTidyCheck.h"
+
+namespace clang {
+namespace tidy {
+namespace hooks {
+
+/// Hook API etxn_reserve takes a parameter specifying the number of
+/// transactions intended to emit from the calling hook. Value of this
+/// parameter is limited, and the function fails if the limit is
+/// exceeded.
+class ReserveLimitCheck : public ClangTidyCheck {
+public:
+  ReserveLimitCheck(StringRef Name, ClangTidyContext *Context)
+      : ClangTidyCheck(Name, Context) {}
+  void registerMatchers(ast_matchers::MatchFinder *Finder) override;
+  void check(const ast_matchers::MatchFinder::MatchResult &Result) override;
+
+private:
+  static const uint64_t MAX_EMIT = 255; // value copied from rippled source
+};
+
+} // namespace hooks
+} // namespace tidy
+} // namespace clang
+
+#endif // LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_HOOKS_RESERVELIMITCHECK_H