+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -124,6 +124,11 @@ New checks
 
   FIXME: add release notes.
 
+- New :doc:`hooks-reserve-limit
+  <clang-tidy/checks/hooks-reserve-limit>` check.
+
+  FIXME: add release notes.
+
 - New :doc:`misc-misleading-identifier <clang-tidy/checks/misc-misleading-identifier>` check.
 
   Reports identifier with unicode right-to-left characters.